continue deflating simulator class
